Handle more errors and correct get listings token api

+++ src/main/java/me/matthew/flink/backpacktfforward/model/BackpackTfApiResponse.java
@@ -134,7 +134,7 @@ public class BackpackTfApiResponse {
         /**
          * Number of keys in the price
          */
-        private Integer keys;
+        private Long keys;
         
         /**
          * Amount of metal in the price
