+++ src/main/java/me/matthew/flink/backpacktfforward/sink/ListingUpsertSink.java
@@ -146,7 +146,7 @@ public class ListingUpsertSink extends RichSinkFunction<ListingUpdate> {
                     else stmt.setNull(7, Types.DOUBLE);
 
                     if (p.getCurrencies() != null && p.getCurrencies().getKeys() != null)
-                        stmt.setInt(8, p.getCurrencies().getKeys());
+                        stmt.setInt(8, p.getCurrencies().getKeys().intValue());
                     else stmt.setNull(8, Types.INTEGER);
 
                     if (p.getValue() != null)
